# apps/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organisation/positions": {
      "post": {
        "operationId": "PositionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Create a position",
        "tags": [
          "Organisation / Positions"
        ]
      },
      "get": {
        "operationId": "PositionsController_search",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by title or description",
            "schema": {
              "minLength": 1,
              "maxLength": 255,
              "example": "Senior Backend Engineer",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sorting field",
            "schema": {
              "default": "id",
              "example": "title",
              "type": "string",
              "enum": [
                "id",
                "title",
                "description",
                "createdAt",
                "updatedAt"
              ]
            }
          },
          {
            "name": "sortDirection",
            "required": false,
            "in": "query",
            "description": "Sorting direction",
            "schema": {
              "default": "asc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Search positions",
        "tags": [
          "Organisation / Positions"
        ]
      }
    },
    "/organisation/positions/{id}": {
      "get": {
        "operationId": "PositionsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Get a position by id",
        "tags": [
          "Organisation / Positions"
        ]
      },
      "patch": {
        "operationId": "PositionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Update a position",
        "tags": [
          "Organisation / Positions"
        ]
      },
      "delete": {
        "operationId": "PositionsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Delete a position",
        "tags": [
          "Organisation / Positions"
        ]
      }
    },
    "/organisation/positions/{id}/subordinates": {
      "post": {
        "operationId": "PositionsController_linkSubordinate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "summary": "Add a subordinate position",
        "tags": [
          "Organisation / Positions"
        ]
      },
      "get": {
        "operationId": "PositionsController_listSubordinates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List subordinate positions",
        "tags": [
          "Organisation / Positions"
        ]
      }
    },
    "/organisation/positions/{id}/subordinates/{subordinateId}": {
      "delete": {
        "operationId": "PositionsController_unlinkSubordinate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subordinateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Delete a subordinate position",
        "tags": [
          "Organisation / Positions"
        ]
      }
    },
    "/organisation/positions/{id}/superiors": {
      "get": {
        "operationId": "PositionsController_listSuperiors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "List superior positions",
        "tags": [
          "Organisation / Positions"
        ]
      }
    },
...
    "/library/questions/{id}/positions/{positionId}": {
      "delete": {
        "operationId": "QuestionsController_detachPosition",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "positionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Invalid data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Detach question from position",
        "tags": [
          "Library / Questions"
        ]
      }
    }
  },
  "info": {
    "title": "Intra API",
    "description": "API documentation for Intra",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreatePositionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Position title",
            "example": "Senior Backend Engineer",
            "maxLength": 255,
            "minLength": 1
          },
          "description": {
            "type": "object",
            "description": "Position description",
            "example": "Backend development, architecture design, code reviews",
            "nullable": true,
            "maxLength": 1000,
            "minLength": 1
          }
        },
        "required": [
          "title"
        ]
      },
      "PositionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "Senior Backend Engineer"
          },
          "description": {
            "type": "object",
            "example": "Backend development, architecture design, code reviews",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-02T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePositionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Position title",
            "example": "Senior Backend Engineer",
            "maxLength": 255,
            "minLength": 1
          },
          "description": {
            "type": "object",
            "description": "Position description",
            "example": "Backend development, architecture design, code reviews",
            "nullable": true,
            "maxLength": 1000,
            "minLength": 1
          }
        }
      },
      "CreatePositionLinkDto": {
        "type": "object",
        "properties": {
          "subordinateId": {
            "type": "number",
            "description": "Id of subordinate position",
            "example": 5
          }
        },
        "required": [
          "subordinateId"
        ]
      },
...
    }
  }
}